server/intelligence/pyproject.toml
[project]
name = "intelligence"
version = "0.1.0"
description = "Intelligence engine for OpenTier"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncpg>=0.30.0",
    "beautifulsoup4>=4.12.0",
    "google-genai>=1.63.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "httpx>=0.28.1",
    "lxml>=5.3.0",
    "markdown>=3.5.0",
    "numpy>=2.4.1",
    "pgvector>=0.4.2",
    "playwright>=1.48.0",
    "protobuf>=6.33.3",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pygithub>=2.1.0",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "sqlalchemy>=2.0.45",
    "tenacity>=8.2.3",
    "torch>=2.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.2",
]

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
